+++ tests/generate_go_ethereum_fixture.py
@@ -292,22 +292,22 @@ def verify_chain_state(web3, chain_data):
 
 
 def mine_transaction_hash(web3, txn_hash):
-    web3.miner.start(1)
+    web3.geth.miner.start(1)
     try:
         return web3.eth.waitForTransactionReceipt(txn_hash, timeout=60)
     finally:
-        web3.miner.stop()
+        web3.geth.miner.stop()
 
 
 def mine_block(web3):
     origin_block_number = web3.eth.blockNumber
 
     start_time = time.time()
-    web3.miner.start(1)
+    web3.geth.miner.start(1)
     while time.time() < start_time + 60:
         block_number = web3.eth.blockNumber
         if block_number > origin_block_number:
-            web3.miner.stop()
+            web3.geth.miner.stop()
             return block_number
         else:
             time.sleep(0.1)
@@ -376,7 +376,7 @@ def setup_chain_state(web3):
     # Block with Transaction
     #
     web3.personal.unlockAccount(coinbase, KEYFILE_PW)
-    web3.miner.start(1)
+    web3.geth.miner.start(1)
     mined_txn_hash = web3.eth.sendTransaction({
         'from': coinbase,
         'to': coinbase,