+++ tests/integration/generate_fixtures/common.py
@@ -194,11 +194,11 @@ def mine_block(web3):
     origin_block_number = web3.eth.blockNumber
 
     start_time = time.time()
-    web3.miner.start(1)
+    web3.geth.miner.start(1)
     while time.time() < start_time + 120:
         block_number = web3.eth.blockNumber
         if block_number > origin_block_number:
-            web3.miner.stop()
+            web3.geth.miner.stop()
             return block_number
         else:
             time.sleep(0.1)
@@ -208,11 +208,11 @@ def mine_block(web3):
 
 def mine_transaction_hash(web3, txn_hash):
     start_time = time.time()
-    web3.miner.start(1)
+    web3.geth.miner.start(1)
     while time.time() < start_time + 120:
         receipt = web3.eth.getTransactionReceipt(txn_hash)
         if receipt is not None:
-            web3.miner.stop()
+            web3.geth.miner.stop()
             return receipt
         else:
             time.sleep(0.1)